Split the release tags into classes that each have separate responsibilities
