+++ lib/dor/release_tags.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Dor
+  module ReleaseTags
+    extend ActiveSupport::Autoload
+
+    eager_autoload do
+      autoload :PurlClient
+      autoload :Purl
+      autoload :IdentityMetadata
+    end
+  end
+end
